+++ src/main/java/com/ordermanagement/item/ItemController.java
@@ -1,4 +1,4 @@
-package com.manageorders.item;
+package com.ordermanagement.item;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
