+++ src/main/java/com/ordermanagement/item/ItemRepository.java
@@ -1,4 +1,4 @@
-package com.manageorders.item;
+package com.ordermanagement.item;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
