+++ src/main/java/com/ordermanagement/stockmovement/IStockMovementRepository.java
@@ -1,4 +1,4 @@
-package com.manageorders.stockmovement;
+package com.ordermanagement.stockmovement;
 
 import org.springframework.data.jpa.repository.JpaRepository;
