+++ src/main/java/com/ordermanagement/stockmovement/IStockMovementService.java
@@ -1,4 +1,4 @@
-package com.manageorders.stockmovement;
+package com.ordermanagement.stockmovement;
 
 import java.util.List;
 import java.util.Optional;
