+++ src/main/java/com/ordermanagement/stockmovement/StockMovementDTO.java
@@ -1,4 +1,4 @@
-package com.manageorders.stockmovement;
+package com.ordermanagement.stockmovement;
 
 import lombok.Data;
