+++ src/main/java/com/ordermanagement/user/UserController.java
@@ -1,4 +1,4 @@
-package com.manageorders.user;
+package com.ordermanagement.user;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
