+++ src/test/java/com/ordermanagement/ManageOrdersApplicationTests.java
@@ -1,4 +1,4 @@
-package com.manageorders;
+package com.ordermanagement;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
